dse: Split code over multiple modules, in subdir
